+++ packages/go_router_linter/lib/src/go_router_lint_plugin.dart
@@ -3,6 +3,7 @@ import 'package:custom_lint_builder/custom_lint_builder.dart';
 
 // 🌎 Project imports:
 import 'package:go_router_linter/src/rules/avoid_hardcoded_routes.dart';
+import 'package:go_router_linter/src/rules/avoid_navigator_named_routes_with_go_router.dart';
 import 'package:go_router_linter/src/rules/missing_go_route_name_property.dart';
 import 'package:go_router_linter/src/rules/use_context_directly_for_go_router.dart';
 
@@ -17,5 +18,6 @@ class _GoRouterLintPlugin extends PluginBase {
     const UseContextDirectlyForGoRouter(),
     const MissingGoRouteNameProperty(),
     const AvoidHardcodedRoutes(),
+    const AvoidNavigatorNamedRoutesWithGoRouter(),
   ];
 }