+++ packages/go_router_linter/test/src/go_router_lint_plugin_test.dart
@@ -30,10 +30,11 @@ void main() {
           ),
         );
         expect(rules, isA<List<LintRule>>());
-        expect(rules, hasLength(3));
+        expect(rules, hasLength(4));
         expect(rules[0], isA<UseContextDirectlyForGoRouter>());
         expect(rules[1], isA<MissingGoRouteNameProperty>());
         expect(rules[2], isA<AvoidHardcodedRoutes>());
+        expect(rules[3], isA<AvoidNavigatorNamedRoutesWithGoRouter>());
       });
     });
   });